+++ src/Loggers/ProcessorFactoryInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * TechDivision\Import\Configuration\Logger\FormatterConfigurationInterface
+ * TechDivision\Import\Loggers\ProcessorFactoryInterface
  *
  * NOTICE OF LICENSE
  *
@@ -12,32 +12,34 @@
  * PHP version 5
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
 
-namespace TechDivision\Import\Configuration\Logger;
+namespace TechDivision\Import\Loggers;
 
-use TechDivision\Import\Configuration\ParamsConfigurationInterface;
+use TechDivision\Import\Configuration\Logger\ProcessorConfigurationInterface;
 
 /**
- * The interface for a handler's formatter configuration.
+ * Interface for processor factory implementations.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
-interface FormatterConfigurationInterface extends ParamsConfigurationInterface
+interface ProcessorFactoryInterface
 {
 
     /**
-     * Return's the formatter's DI ID to use.
+     * Creates a new processor instance based on the passed configuration.
      *
-     * @return string The type
+     * @param \TechDivision\Import\Configuration\Logger\ProcessorConfigurationInterface $processorConfiguration The processor configuration
+     *
+     * @return object The processor instance
      */
-    public function getId();
+    public function factory(ProcessorConfigurationInterface $processorConfiguration);
 }