+++ src/Utils/LoggerFactory.php
@@ -26,11 +26,13 @@ use TechDivision\Import\Configuration\LoggerConfigurationInterface;
 /**
  * Logger factory implementation.
  *
- * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link      https://github.com/techdivision/import
- * @link      http://www.techdivision.com
+ * @author     Tim Wagner <t.wagner@techdivision.com>
+ * @copyright  2016 TechDivision GmbH <info@techdivision.com>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       https://github.com/techdivision/import
+ * @link       http://www.techdivision.com
+ * @deprecated Since 15.0.0
+ * @see        TechDivision\Import\Loggers\MonologLoggerFactory
  */
 class LoggerFactory
 {